fix(new-address-modal): rename
OpenNewAddressModal renamed to NewAddressButton
NewAddressButton component matches convention.
Added NewAddressButton to Taro Actions Tab
