+++ src/components/designer/taro/actions/index.ts
@@ -3,4 +3,4 @@ export { default as OpenSendAssetModal } from './OpenSendAssetModal';
 export { default as MintAssetModal } from './MintAssetModal';
 export { default as SendAssetModal } from './SendAssetModal';
 export { default as NewAddressModal } from './NewAddressModal';
-export { default as OpenNewAddressModal } from './OpenNewAddressModal';
+export { default as NewAddressButton } from './NewAddressButton';
